pyproject.toml
[project]
name = "fleetmdm"
version = "0.1.0"
description = "Local-first FleetMDM controller: inventory ingest, policy checks, and compliance reporting."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Sarvesh Kapre" }]
keywords = ["mdm", "fleet", "compliance", "inventory"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "ruff>=0.4.0",
  "pyright>=1.1.360",
  "build>=1.2.1",
  "pip-audit>=2.7.0",
  "bandit>=1.7.8",
]

[project.scripts]
fleetmdm = "fleetmdm.cli:app"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]
ignore = ["UP017"]

[tool.ruff.lint.isort]
known-first-party = ["fleetmdm"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownParameterType = false
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
addopts = "-q"
